python/requirements.txt
pylint
kafka-python
requests